factored out automation bits to own assembly
