+++ source/DasBlog All.sln
@@ -53,6 +53,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DasBlog.Tests.SmokeTest", "
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DasBlog.Tests.FunctionalTests", "DasBlog.Tests\FunctionalTests\DasBlog.Tests.FunctionalTests.csproj", "{42BAE6DE-FDF7-497E-B08C-4B9035C7F1FF}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DasBlog.Tests.Automation", "DasBlog.Tests\Automation\DasBlog.Tests.Automation\DasBlog.Tests.Automation.csproj", "{85B28FFC-EBB9-4E18-A1F4-9C352789725F}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug - All|Any CPU = Debug - All|Any CPU
@@ -184,6 +186,12 @@ Global
 		{42BAE6DE-FDF7-497E-B08C-4B9035C7F1FF}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{42BAE6DE-FDF7-497E-B08C-4B9035C7F1FF}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{42BAE6DE-FDF7-497E-B08C-4B9035C7F1FF}.Release|Any CPU.Build.0 = Release|Any CPU
+		{85B28FFC-EBB9-4E18-A1F4-9C352789725F}.Debug - All|Any CPU.ActiveCfg = Debug|Any CPU
+		{85B28FFC-EBB9-4E18-A1F4-9C352789725F}.Debug - All|Any CPU.Build.0 = Debug|Any CPU
+		{85B28FFC-EBB9-4E18-A1F4-9C352789725F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{85B28FFC-EBB9-4E18-A1F4-9C352789725F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{85B28FFC-EBB9-4E18-A1F4-9C352789725F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{85B28FFC-EBB9-4E18-A1F4-9C352789725F}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -203,6 +211,7 @@ Global
 		{DD4314EB-0163-47B6-A42B-5C47A0DEFBCB} = {F0F076E7-9863-46B7-8C3C-5C220D4BC118}
 		{425AB01F-C15E-4D65-A40A-69CC69D5A396} = {F0F076E7-9863-46B7-8C3C-5C220D4BC118}
 		{42BAE6DE-FDF7-497E-B08C-4B9035C7F1FF} = {F0F076E7-9863-46B7-8C3C-5C220D4BC118}
+		{85B28FFC-EBB9-4E18-A1F4-9C352789725F} = {F0F076E7-9863-46B7-8C3C-5C220D4BC118}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {8F3ACFB8-CB52-4953-A1CF-4151E2789889}
