+++ source/DasBlog.Tests/Automation/DasBlog.Tests.Automation/Common/Constants.cs
@@ -1,4 +1,4 @@
-﻿namespace DasBlog.Tests.SmokeTest.Common
+﻿namespace DasBlog.Tests.Automation.Common
 {
 	public class Constants
 	{
