+++ source/DasBlog.Tests/Automation/DasBlog.Tests.Automation/Dom/ActivityPage.cs
@@ -1,0 +1,13 @@
+﻿using DasBlog.Tests.Automation.Common;
+using DasBlog.Tests.Automation.Selenium.Interfaces;
+
+namespace DasBlog.Tests.Automation.Dom
+{
+	public class ActivityPage : Page
+	{
+		public ActivityPage(IBrowser browser) : base(browser, Constants.ActivityPage)
+		{
+
+		}
+	}
+}
