+++ source/DasBlog.Tests/Automation/DasBlog.Tests.Automation/Dom/HomePage.cs
@@ -1,8 +1,8 @@
 ﻿using WebAppConstants = DasBlog.Core.Common.Constants;
-using DasBlog.Tests.SmokeTest.Common;
-using DasBlog.Tests.SmokeTest.Selenium.Interfaces;
+using DasBlog.Tests.Automation.Common;
+using DasBlog.Tests.Automation.Selenium.Interfaces;
 
-namespace DasBlog.Tests.SmokeTest.Dom
+namespace DasBlog.Tests.Automation.Dom
 {
 	public class HomePage : Page
 	{
