+++ source/DasBlog.Tests/Automation/DasBlog.Tests.Automation/Dom/Http404Page.cs
@@ -1,6 +1,6 @@
-﻿using DasBlog.Tests.SmokeTest.Selenium.Interfaces;
+﻿using DasBlog.Tests.Automation.Selenium.Interfaces;
 
-namespace DasBlog.Tests.SmokeTest.Dom
+namespace DasBlog.Tests.Automation.Dom
 {
 	public class Http404Page :  Page
 	{
