+++ source/DasBlog.Tests/Automation/DasBlog.Tests.Automation/Dom/LoginPage.cs
@@ -1,8 +1,8 @@
-﻿using DasBlog.Tests.SmokeTest.Common;
-using DasBlog.Tests.SmokeTest.Selenium.Interfaces;
+﻿using DasBlog.Tests.Automation.Common;
+using DasBlog.Tests.Automation.Selenium.Interfaces;
 using AppConstants = DasBlog.Core.Common.Constants;
 
-namespace DasBlog.Tests.SmokeTest.Dom
+namespace DasBlog.Tests.Automation.Dom
 {
 	public class LoginPage : Page
 	{
