+++ source/DasBlog.Tests/Automation/DasBlog.Tests.Automation/Dom/SubscribePage.cs
@@ -1,7 +1,7 @@
-﻿using DasBlog.Tests.SmokeTest.Common;
-using DasBlog.Tests.SmokeTest.Selenium.Interfaces;
+﻿using DasBlog.Tests.Automation.Common;
+using DasBlog.Tests.Automation.Selenium.Interfaces;
 
-namespace DasBlog.Tests.SmokeTest.Dom
+namespace DasBlog.Tests.Automation.Dom
 {
 	public class SubscribePage : Page
 	{
