+++ source/DasBlog.Tests/Automation/DasBlog.Tests.Automation/Dom/UsersPage.cs
@@ -1,0 +1,12 @@
+﻿using DasBlog.Tests.Automation.Common;
+using DasBlog.Tests.Automation.Selenium.Interfaces;
+
+namespace DasBlog.Tests.Automation.Dom
+{
+	public class UsersPage : Page
+	{
+		public UsersPage(IBrowser browser) : base(browser, Constants.UsersPage)
+		{
+		}
+	}
+}
