--- source/DasBlog.Tests/SmokeTest/Dom/ActivityPage.cs
@@ -1,13 +1,0 @@
-﻿using DasBlog.Tests.SmokeTest.Common;
-using DasBlog.Tests.SmokeTest.Selenium.Interfaces;
-
-namespace DasBlog.Tests.SmokeTest.Dom
-{
-	public class ActivityPage : Page
-	{
-		public ActivityPage(IBrowser browser) : base(browser, Constants.ActivityPage)
-		{
-
-		}
-	}
-}
