--- source/DasBlog.Tests/SmokeTest/Dom/UsersPage.cs
@@ -1,12 +1,0 @@
-﻿using DasBlog.Tests.SmokeTest.Common;
-using DasBlog.Tests.SmokeTest.Selenium.Interfaces;
-
-namespace DasBlog.Tests.SmokeTest.Dom
-{
-	public class UsersPage : Page
-	{
-		public UsersPage(IBrowser browser) : base(browser, Constants.UsersPage)
-		{
-		}
-	}
-}
