+++ source/DasBlog.Tests/SmokeTest/Smoking/Interfaces/ITestExecutor.cs
@@ -1,6 +1,5 @@
 ﻿using System.Collections.Generic;
 using System.Runtime.CompilerServices;
-using DasBlog.Tests.SmokeTest.Smoking;
 
 namespace DasBlog.Tests.SmokeTest.Smoking.Interfaces
 {
