+++ source/DasBlog.Tests/SmokeTest/Support/DasBlogInstallationOptions.cs
